python_services/requirements.txt
# 短视频创作自动化API服务 - Python依赖

# ==================== Web框架 ====================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ==================== 数据验证 ====================
pydantic==2.5.3
pydantic-settings==2.1.0

# ==================== 认证与安全 ====================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# ==================== 数据库 ====================
sqlalchemy==2.0.25
alembic==1.13.1

# ==================== 任务队列（可选） ====================
celery==5.3.4
redis==5.0.1

# ==================== HTTP客户端 ====================
httpx==0.26.0
requests==2.31.0

# ==================== 日志 ====================
loguru==0.7.2

# ==================== 工具库 ====================
aiofiles==23.2.1
python-dateutil==2.8.2

# ==================== 阿里云SDK ====================
dashscope==1.14.0
oss2==2.18.4

# ==================== OpenAI兼容SDK ====================
openai==1.10.0

# ==================== 数据处理 ====================
pandas==2.2.0
numpy==1.26.3

# ==================== 文件处理 ====================
Pillow==10.2.0

# ==================== 测试（开发时） ====================
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
